+++ frontend/src/App.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 import Header from '~/components/common/Header/Header';
-import PageTemplate from '~/components/pages/PageTemplate/PageTemplate';
+import PageTemplate from '~/pages/PageTemplate/PageTemplate';
 
 const App = () => {
   return (
