+++ frontend/src/components/common/DateCell/DateCell.styled.ts
@@ -1,5 +1,5 @@
 import { css, styled } from 'styled-components';
-import { DateCellProps } from '~/components/DateCell/DateCell';
+import { DateCellProps } from '~/components/common/DateCell/DateCell';
 import type { DateCellSize } from '~/types/size';
 
 interface WrapperProps {
